Clean up type resolvers
