+++ .editorconfig
@@ -8,6 +8,7 @@ trim_trailing_whitespace = true
 [*.md]
 trim_trailing_whitespace = false
 indent_style = space
+indent_size = 2
 
 [output.txt]
 insert_final_newline = false
