+++ ARCH_REFACTOR.md
@@ -1,6 +1,6 @@
 # Architectural refactoring plan
 
-- [ ] ARCH-001: [P1] Type resolution is carrying too many architectural roles. typeResolver.ts (line 33) handles cache policy, cycle detection, warning buffering, external-type policy, TS internals, and every AnyType construction path. The long ordered branch chain around type flags (line 168) makes new type support risky because behavior depends on branch order. A TypeResolutionPipeline pattern would help: small resolvers like UnionResolver, ObjectResolver, ExternalResolver, ConditionalResolver, coordinated by one session object that owns cache, cycle guards, and diagnostics.
+- [x] ARCH-001: [P1] Type resolution is carrying too many architectural roles. typeResolver.ts (line 33) handles cache policy, cycle detection, warning buffering, external-type policy, TS internals, and every AnyType construction path. The long ordered branch chain around type flags (line 168) makes new type support risky because behavior depends on branch order. A TypeResolutionPipeline pattern would help: small resolvers like UnionResolver, ObjectResolver, ExternalResolver, ConditionalResolver, coordinated by one session object that owns cache, cycle guards, and diagnostics.
 - [ ] ARCH-002: [P1] ParserContext is a mutable ambient state bag. It exposes stacks and substitutions directly in parser.ts (line 96), then many parsers manually push/pop nodes and symbols. Warning locations later depend on that stack in typeResolver.ts (line 626). Introduce scoped helpers like context.withSymbol(...), context.withSourceNode(...), and context.withTypeSubstitutions(...); that would make stack correctness local and remove a lot of fragile finally boilerplate.
 - [ ] ARCH-003: [P1] Export parsing has become a second orchestration hub. exportParser.ts (line 19) does export-symbol targeting, namespace merging, re-export metadata, type acquisition, declaration dispatch, extends extraction, documentation, warning emission, and model construction. A stronger pattern would be: first normalize exports into an ExportDescriptor, then convert descriptors into ExportNodes. That would also remove repeated “append namespace members” paths and make alias/re-export behavior testable in isolation.
 - [ ] ARCH-004: [P2] Function and class signature parsing are duplicated and already diverge. functionParser.ts (line 69) and classParser.ts (line 257) both parse parameters, docs, visibility, optionality, defaults, and return types. They differ on non-literal default values: class parameters keep initializer.getText() while function parameters do not. Expand signatureParser.ts into the single owner of parseCallSignature, parseParameter, parseReturnType, and parameter docs.
