+++ src/parsers/typeResolver.ts
@@ -1,32 +1,14 @@
 import ts from 'typescript';
-import { getDocumentationFromSymbol } from './documentationParser';
-import { ParserContext } from '../parser';
-import { parseClassType } from './classParser';
-import { parseFunctionType } from './functionParser';
-import { parseObjectType } from './objectParser';
-import { parseEnum } from './enumParser';
-import {
-	ObjectNode,
-	TypeParameterNode,
-	ArrayNode,
-	ExternalTypeNode,
-	IntrinsicNode,
-	UnionNode,
-	TupleNode,
-	LiteralNode,
-	IntersectionNode,
-	FunctionNode,
-	AnyType,
-} from '../models';
-import { resolveUnionType } from './unionTypeResolver';
-import { getFullName, isInternalSymbolName } from './common';
-import { TypeName } from '../models/typeName';
-
-const unsupportedFallbackTypes = new WeakSet<ts.Type>();
+import { type AnyType } from '../models';
+import { type ParserContext } from '../parser';
+import { TypeResolutionSession } from './typeResolutionSession';
 
 /**
+ * Public facade for type resolution. The implementation is intentionally kept
+ * in TypeResolutionSession and resolver modules so this entry point remains the
+ * stable API used by the rest of the parser.
  *
- * @param type TypeScript type to resolve
+ * @param type TypeScript type to resolve.
  * @param typeNode TypeScript TypeNode associated with the type, if available. It can be used to preserve the authored type name.
  * @param context Parser context containing TypeScript checker and other utilities.
  */
@@ -35,789 +17,5 @@ export function resolveType(
 	typeNode: ts.TypeNode | undefined,
 	context: ParserContext,
 ): AnyType {
-	const { resolvedTypeCache, typeStack, typeParameterSubstitutions } = context;
-
-	const typeId = getTypeId(type);
-	const hasTypeParameterSubstitutions = Boolean(typeParameterSubstitutions?.size);
-
-	// Build a cache key that incorporates both the type identity and the current
-	// stack depth, because shouldResolveObject / shouldInclude are depth-sensitive.
-	const cacheKey = typeId !== undefined ? `${typeId}@${typeStack.length}` : undefined;
-
-	// Check the cache first for types we've already resolved.
-	// Only use cache when there's no typeNode (which can affect alias resolution)
-	// and when we're not in a recursive context (type isn't already on the stack).
-	if (
-		cacheKey !== undefined &&
-		!typeNode &&
-		!hasTypeParameterSubstitutions &&
-		resolvedTypeCache.has(cacheKey) &&
-		!unsupportedFallbackTypes.has(type) &&
-		!typeStack.includes(typeId!)
-	) {
-		return resolvedTypeCache.get(cacheKey)!;
-	}
-
-	const result = resolveTypeUncached(type, typeNode, context);
-
-	// Cache the result for future lookups when there's no typeNode influence
-	if (
-		cacheKey !== undefined &&
-		!typeNode &&
-		!hasTypeParameterSubstitutions &&
-		!unsupportedFallbackTypes.has(type) &&
-		!typeStack.includes(typeId!)
-	) {
-		resolvedTypeCache.set(cacheKey, result);
-	}
-
-	return result;
-}
-
-function resolveTypeUncached(
-	type: ts.Type,
-	typeNode: ts.TypeNode | undefined,
-	context: ParserContext,
-): AnyType {
-	const { checker, typeStack, includeExternalTypes } = context;
-
-	const typeId = getTypeId(type);
-
-	// If the typeStack contains type.id we're dealing with an object that references itself.
-	// To prevent getting stuck in an infinite loop we just set it to an objectNode
-	// However, we should not apply this check to intrinsic types like any/unknown/string/etc
-	// as they can appear multiple times without causing infinite recursion
-	const isIntrinsicType =
-		(type.flags &
-			(ts.TypeFlags.Any |
-				ts.TypeFlags.Unknown |
-				ts.TypeFlags.String |
-				ts.TypeFlags.Number |
-				ts.TypeFlags.Boolean |
-				ts.TypeFlags.ESSymbol |
-				ts.TypeFlags.UniqueESSymbol |
-				ts.TypeFlags.Never |
-				ts.TypeFlags.Undefined |
-				ts.TypeFlags.Null |
-				ts.TypeFlags.Void |
-				ts.TypeFlags.BigInt)) !==
-		0;
-
-	// Check for cycles before pushing to stack
-	// If we're already resolving this type, return a shallow object with type info but no properties
-	const shouldDetectCycles = !isIntrinsicType && typeId !== undefined;
-	const isAlreadyOnStack = shouldDetectCycles && typeStack.includes(typeId);
-
-	// Push type to stack BEFORE calling getFullName to catch cycles that occur
-	// when getFullName resolves generic type arguments that may reference back to this type.
-	// We track whether we pushed so we can correctly pop in the finally block.
-	const shouldPushToStack = shouldDetectCycles && !isAlreadyOnStack;
-	if (shouldPushToStack) {
-		typeStack.push(typeId);
-	}
-
-	let typeName: TypeName | undefined;
-
-	try {
-		// SubstitutionTypes are checker-internal placeholders created while TypeScript
-		// evaluates conditional/infer/mapped types. They represent "baseType, but
-		// under this constraint" rather than a syntax form we can emit directly.
-		// Try this before getFullName(), which may resolve type arguments and emit
-		// warnings that would be misleading if the substitution fallback succeeds.
-		if (!isAlreadyOnStack && hasExactFlag(type, ts.TypeFlags.Substitution)) {
-			const substitutionFallback = resolveSubstitutionFallback(type, context);
-			if (substitutionFallback) {
-				return substitutionFallback;
-			}
-		}
-
-		const nameResolutionWarnings: Array<Parameters<ParserContext['onWarning']>[0]> = [];
-		if (hasExactFlag(type, ts.TypeFlags.Conditional)) {
-			const onWarning = context.onWarning;
-			context.onWarning = (warning) => {
-				nameResolutionWarnings.push(warning);
-			};
-			try {
-				typeName = getFullName(type, typeNode, context);
-			} finally {
-				context.onWarning = onWarning;
-			}
-		} else {
-			typeName = getFullName(type, typeNode, context);
-		}
-
-		const replayNameResolutionWarnings = () => {
-			for (const warning of nameResolutionWarnings) {
-				context.onWarning(warning);
-			}
-			nameResolutionWarnings.length = 0;
-		};
-
-		// Conditional type name resolution may emit warnings for type arguments.
-		// Replay them only on return paths that keep the computed typeName.
-		const withNameResolutionWarnings = <T extends AnyType>(resolvedType: T): T => {
-			replayNameResolutionWarnings();
-			return resolvedType;
-		};
-
-		// If this type was already on the stack, return a shallow version with type info but no properties
-		if (isAlreadyOnStack) {
-			return withNameResolutionWarnings(createShallowType(type, typeName, checker));
-		}
-
-		if (hasExactFlag(type, ts.TypeFlags.TypeParameter) && type.symbol) {
-			const substitution = context.typeParameterSubstitutions?.get(type.symbol);
-			if (
-				substitution &&
-				substitution !== type &&
-				(!(substitution.flags & ts.TypeFlags.TypeParameter) || substitution.symbol !== type.symbol)
-			) {
-				return resolveType(substitution, undefined, context);
-			}
-
-			// If we have a typeNode, check if it resolves to a more concrete type than the TypeParameter.
-			// This handles cases where TypeScript doesn't fully instantiate generic parameters,
-			// but the typeNode (authored code) references the actual concrete type.
-			if (typeNode && ts.isTypeReferenceNode(typeNode)) {
-				// Get the symbol that the type reference points to
-				const symbol = checker.getSymbolAtLocation(typeNode.typeName);
-				if (symbol && !(symbol.flags & ts.SymbolFlags.TypeParameter)) {
-					// The symbol is not a type parameter - it's a concrete type alias or interface
-					// Get the type from the symbol's declaration
-					const symbolType = checker.getDeclaredTypeOfSymbol(symbol);
-					if (symbolType && !hasExactFlag(symbolType, ts.TypeFlags.TypeParameter)) {
-						return withNameResolutionWarnings(resolveType(symbolType, typeNode, context));
-					}
-				}
-			}
-
-			const declaration = type.symbol.declarations?.[0] as ts.TypeParameterDeclaration | undefined;
-			const constraintType = declaration?.constraint
-				? checker.getBaseConstraintOfType(type)
-				: undefined;
-
-			return withNameResolutionWarnings(
-				new TypeParameterNode(
-					type.symbol.name,
-					constraintType ? resolveType(constraintType, undefined, context) : undefined,
-					declaration?.default
-						? resolveType(checker.getTypeAtLocation(declaration.default), undefined, context)
-						: undefined,
-				),
-			);
-		}
-
-		if (checker.isArrayType(type)) {
-			// @ts-expect-error - Private method
-			const arrayType: ts.Type = checker.getElementTypeOfArrayType(type);
-			return withNameResolutionWarnings(
-				new ArrayNode(
-					type.aliasSymbol?.name
-						? new TypeName(type.aliasSymbol?.name, typeName?.namespaces, typeName?.typeArguments)
-						: undefined,
-					resolveType(arrayType, undefined, context),
-				),
-			);
-		}
-
-		if (!includeExternalTypes && isTypeExternal(type, checker)) {
-			// Determine the best name to use for this external type.
-			// When a type fully resolves to an external interface (e.g., `Event` from lib.dom.d.ts),
-			// TypeScript provides no aliasSymbol - the type is just the resolved interface.
-			// In this case, use the resolved symbol's name (e.g., `Event`, `KeyboardEvent`).
-			//
-			// When a type is an external alias (e.g., `Point` from a package that wraps `{ x, y }`),
-			// TypeScript preserves the aliasSymbol. Use the alias name to preserve the author's intent.
-			//
-			// The key insight: if there's no aliasSymbol but there is a resolved symbol from
-			// node_modules, TypeScript has fully resolved the type and we should use that name.
-			const resolvedSymbol = type.getSymbol();
-			const resolvedSymbolName = resolvedSymbol?.getName?.();
-
-			let externalTypeName: string | undefined;
-			// If the resolved symbol is external and is a named interface (not anonymous internal names),
-			// and there's no local alias wrapping it, use the resolved interface name.
-			const resolvedIsExternalInterface =
-				resolvedSymbolName &&
-				!isInternalSymbolName(resolvedSymbolName) &&
-				isSymbolExternal(resolvedSymbol, checker, false) &&
-				(resolvedSymbol?.flags ?? 0) & ts.SymbolFlags.Interface;
-
-			if (resolvedIsExternalInterface && !type.aliasSymbol) {
-				// Fully resolved external interface (e.g., Event, KeyboardEvent)
-				externalTypeName = resolvedSymbolName;
-			} else {
-				// Use the authored alias name, falling back to resolved symbol name
-				externalTypeName = typeName?.name || type.aliasSymbol?.getName?.() || resolvedSymbolName;
-			}
-
-			if (!externalTypeName) {
-				return withNameResolutionWarnings(new IntrinsicNode('any'));
-			}
-
-			// Fixes a weird TS behavior where it doesn't show the alias name but resolves to the actual type in case of RefCallback.
-			if (externalTypeName === 'bivarianceHack') {
-				return withNameResolutionWarnings(
-					new ExternalTypeNode(new TypeName('RefCallback', ['React'], typeName?.typeArguments)),
-				);
-			}
-
-			return withNameResolutionWarnings(
-				new ExternalTypeNode(
-					new TypeName(externalTypeName, typeName?.namespaces, typeName?.typeArguments),
-				),
-			);
-		}
-
-		if (hasExactFlag(type, ts.TypeFlags.Boolean)) {
-			return withNameResolutionWarnings(new IntrinsicNode('boolean'));
-		}
-
-		if (hasExactFlag(type, ts.TypeFlags.Void)) {
-			return withNameResolutionWarnings(new IntrinsicNode('void'));
-		}
-
-		if (includesCompositeFlag(type, ts.TypeFlags.EnumLike)) {
-			let symbol = type.aliasSymbol ?? type.getSymbol();
-			if ('value' in type) {
-				// weird edge case - when an enum has one member only, type.getSymbol() returns the symbol of the member
-				symbol = symbol?.parent;
-			}
-
-			if (!symbol) {
-				return withNameResolutionWarnings(new IntrinsicNode('any'));
-			}
-
-			return withNameResolutionWarnings(parseEnum(symbol, context));
-		}
-
-		if (type.isUnion()) {
-			return withNameResolutionWarnings(resolveUnionType(type, typeName, typeNode, context));
-		}
-
-		if (type.isIntersection()) {
-			const memberTypes: AnyType[] = [];
-
-			for (const memberType of type.types) {
-				memberTypes.push(resolveType(memberType, undefined, context));
-			}
-
-			if (memberTypes.length === 0) {
-				throw new Error('Encountered an intersection type with no members');
-			}
-
-			if (memberTypes.length === 1) {
-				return withNameResolutionWarnings(memberTypes[0]);
-			}
-
-			if (memberTypes.length > 1) {
-				const callSignatures = type.getCallSignatures();
-				if (callSignatures.length >= 1) {
-					return withNameResolutionWarnings(parseFunctionType(type, context)!);
-				}
-
-				const objectType = parseObjectType(type, typeName, context, resolveType);
-				if (objectType) {
-					return withNameResolutionWarnings(
-						new IntersectionNode(typeName, memberTypes, objectType.properties),
-					);
-				}
-
-				return withNameResolutionWarnings(new IntersectionNode(typeName, memberTypes, []));
-			}
-		}
-
-		if (checker.isTupleType(type)) {
-			return withNameResolutionWarnings(
-				new TupleNode(
-					typeName,
-					(type as ts.TupleType).typeArguments?.map((x) => resolveType(x, undefined, context)) ??
-						[],
-				),
-			);
-		}
-
-		if (hasExactFlag(type, ts.TypeFlags.String)) {
-			return withNameResolutionWarnings(new IntrinsicNode('string'));
-		}
-
-		if (hasExactFlag(type, ts.TypeFlags.Number)) {
-			return withNameResolutionWarnings(new IntrinsicNode('number'));
-		}
-
-		if (hasExactFlag(type, ts.TypeFlags.BigInt)) {
-			return withNameResolutionWarnings(new IntrinsicNode('bigint'));
-		}
-
-		if (
-			hasExactFlag(type, ts.TypeFlags.ESSymbol) ||
-			hasExactFlag(type, ts.TypeFlags.UniqueESSymbol)
-		) {
-			return withNameResolutionWarnings(new IntrinsicNode('symbol', typeName));
-		}
-
-		if (hasExactFlag(type, ts.TypeFlags.Undefined)) {
-			return withNameResolutionWarnings(new IntrinsicNode('undefined'));
-		}
-
-		if (hasExactFlag(type, ts.TypeFlags.Any)) {
-			// Special case: if the authored typeNode is a union (e.g., `AliasedAny | undefined`),
-			// we should resolve it as a union to preserve alias information,
-			// even though TypeScript simplifies `any | T` to just `any` in the type system.
-			if (typeNode && ts.isUnionTypeNode(typeNode)) {
-				const unionTypes: AnyType[] = [];
-				for (const memberTypeNode of typeNode.types) {
-					// Get the type from the type node
-					const memberType = checker.getTypeFromTypeNode(memberTypeNode);
-					// Recursively resolve, passing the memberTypeNode so alias information is preserved
-					unionTypes.push(resolveType(memberType, memberTypeNode, context));
-				}
-				return withNameResolutionWarnings(new UnionNode(typeName, unionTypes));
-			}
-			return withNameResolutionWarnings(new IntrinsicNode('any', typeName));
-		}
-
-		if (hasExactFlag(type, ts.TypeFlags.Unknown)) {
-			return withNameResolutionWarnings(new IntrinsicNode('unknown', typeName));
-		}
-
-		if (includesCompositeFlag(type, ts.TypeFlags.Literal)) {
-			if (type.isLiteral()) {
-				return withNameResolutionWarnings(
-					new LiteralNode(
-						type.isStringLiteral() ? `"${type.value}"` : type.value,
-						typeName,
-						getDocumentationFromSymbol(type.symbol, checker),
-					),
-				);
-			}
-
-			return withNameResolutionWarnings(new LiteralNode(checker.typeToString(type)));
-		}
-
-		if (hasExactFlag(type, ts.TypeFlags.Null)) {
-			return withNameResolutionWarnings(new IntrinsicNode('null'));
-		}
-
-		// TODO: currently types can be either a "function" or an "object" but not both.
-		// In reality, type can have both call signatures and properties.
-		// Consider creating a new type that can handle both.
-		const callSignatures = type.getCallSignatures();
-		if (callSignatures.length >= 1) {
-			return withNameResolutionWarnings(parseFunctionType(type, context)!);
-		}
-
-		// Check for class types (have construct signatures)
-		const constructSignatures = type.getConstructSignatures();
-		if (constructSignatures.length >= 1) {
-			const classType = parseClassType(type, context);
-			if (classType) {
-				return withNameResolutionWarnings(classType);
-			}
-		}
-
-		const objectType = parseObjectType(type, typeName, context, resolveType);
-		if (objectType) {
-			return withNameResolutionWarnings(objectType);
-		}
-
-		// Object without properties or object keyword
-		if (
-			hasExactFlag(type, ts.TypeFlags.Object) ||
-			(hasExactFlag(type, ts.TypeFlags.NonPrimitive) && checker.typeToString(type) === 'object')
-		) {
-			return withNameResolutionWarnings(new ObjectNode(typeName, [], undefined));
-		}
-
-		if (hasExactFlag(type, ts.TypeFlags.Never)) {
-			return withNameResolutionWarnings(new IntrinsicNode('never', typeName));
-		}
-
-		if (hasExactFlag(type, ts.TypeFlags.Conditional)) {
-			const conditionalType = type as ts.ConditionalType;
-			if (conditionalType.resolvedTrueType && conditionalType.resolvedFalseType) {
-				return new UnionNode(undefined, [
-					// TODO: Pass TypeNode here to resolve aliases correctly
-					resolveType((type as ts.ConditionalType).resolvedTrueType!, undefined, context),
-					resolveType((type as ts.ConditionalType).resolvedFalseType!, undefined, context),
-				]);
-			} else if (conditionalType.resolvedTrueType) {
-				return resolveType(conditionalType.resolvedTrueType, undefined, context);
-			} else if (conditionalType.resolvedFalseType) {
-				return resolveType(conditionalType.resolvedFalseType, undefined, context);
-			}
-		}
-
-		// Index (keyof T) and IndexedAccess (T[K]) types can't be represented directly.
-		// Expand them via getBaseConstraintOfType to a representable form.
-		// When the base constraint is unavailable (e.g., T[K] with unresolved type parameters),
-		// fall back to 'any' silently — this is an expected limitation, not a parser bug.
-		if (hasExactFlag(type, ts.TypeFlags.Index) || hasExactFlag(type, ts.TypeFlags.IndexedAccess)) {
-			const baseConstraint = checker.getBaseConstraintOfType(type);
-			if (baseConstraint) {
-				return resolveType(baseConstraint, undefined, context);
-			}
-			return new IntrinsicNode('any', typeName);
-		}
-
-		unsupportedFallbackTypes.add(type);
-		for (const warning of nameResolutionWarnings) {
-			context.onWarning(warning);
-		}
-		reportUnsupportedTypeFallback(type, typeNode, context);
-
-		return new IntrinsicNode('any', typeName);
-	} finally {
-		// Only pop if we actually pushed
-		if (shouldPushToStack) {
-			typeStack.pop();
-		}
-	}
-}
-
-function resolveSubstitutionFallback(type: ts.Type, context: ParserContext): AnyType | undefined {
-	const substitutionType = type as ts.SubstitutionType;
-
-	return (
-		resolveSubstitutionCandidate(substitutionType.baseType, context) ??
-		resolveSubstitutionCandidate(substitutionType.constraint, context)
-	);
-}
-
-function resolveSubstitutionCandidate(
-	candidateType: ts.Type,
-	context: ParserContext,
-): AnyType | undefined {
-	if (hasExactFlag(candidateType, ts.TypeFlags.Substitution)) {
-		return undefined;
-	}
-
-	// This is a probe, not the real fallback path. If the candidate still needs
-	// an unsupported-type warning, reject it and let the original substitution
-	// report a single warning with the best source location.
-	const warnings: unknown[] = [];
-	const onWarning = context.onWarning;
-	context.onWarning = (warning) => {
-		warnings.push(warning);
-	};
-
-	try {
-		const resolvedCandidate = resolveType(candidateType, undefined, context);
-		return warnings.length === 0 && !isAnyNode(resolvedCandidate) ? resolvedCandidate : undefined;
-	} finally {
-		context.onWarning = onWarning;
-	}
-}
-
-function isAnyNode(typeNode: AnyType): boolean {
-	return typeNode instanceof IntrinsicNode && typeNode.intrinsic === 'any' && !typeNode.typeName;
-}
-
-const allowedBuiltInTsTypes = new Set([
-	'Pick',
-	'Omit',
-	'ReturnType',
-	'Parameters',
-	'InstanceType',
-	'Partial',
-	'Required',
-	'Readonly',
-	'Exclude',
-	'Extract',
-]);
-
-const allowedBuiltInReactTypes = new Set([
-	'React.MemoExoticComponent',
-	'React.NamedExoticComponent',
-	'React.FC',
-	'React.FunctionComponent',
-	'React.ForwardRefExoticComponent',
-]);
-
-/**
- * Checks if a symbol is defined externally (in node_modules), excluding allowed built-in types.
- */
-function isSymbolExternal(
-	symbol: ts.Symbol | undefined,
-	checker: ts.TypeChecker,
-	checkAllowList: boolean = true,
-): boolean {
-	if (!symbol) return false;
-	return (
-		symbol.declarations?.some((x) => {
-			const sourceFileName = x.getSourceFile().fileName;
-			const definedExternally = sourceFileName.includes('node_modules');
-			if (!definedExternally) return false;
-			if (!checkAllowList) return true;
-			return !(
-				(allowedBuiltInTsTypes.has(checker.getFullyQualifiedName(symbol)) &&
-					/node_modules\/typescript\/lib/.test(sourceFileName)) ||
-				(allowedBuiltInReactTypes.has(checker.getFullyQualifiedName(symbol)) &&
-					/node_modules\/@types\/react/.test(sourceFileName))
-			);
-		}) ?? false
-	);
-}
-
-function isTypeExternal(type: ts.Type, checker: ts.TypeChecker): boolean {
-	const symbol = type.aliasSymbol ?? type.getSymbol();
-	return isSymbolExternal(symbol, checker);
-}
-
-function hasExactFlag(type: ts.Type, flag: number) {
-	return (type.flags & flag) === flag;
-}
-
-function includesCompositeFlag(type: ts.Type, flag: number) {
-	return (type.flags & flag) !== 0;
-}
-
-function reportUnsupportedTypeFallback(
-	type: ts.Type,
-	typeNode: ts.TypeNode | undefined,
-	context: ParserContext,
-) {
-	const location = getWarningLocation(type, typeNode, context);
-	const typeFlags = getTypeFlagNames(type.flags);
-	const formattedTypeFlags = typeFlags.join(' | ');
-	const typeText = getWarningTypeText(type, context);
-	const sourceText = getWarningSourceText(location.node);
-	const resolvingText =
-		sourceText && sourceText !== typeText ? ` while resolving "${sourceText}"` : '';
-
-	context.onWarning({
-		code: 'unsupported-type-fallback',
-		message: `Type extraction warning: Unable to handle type "${typeText}" with flag "${formattedTypeFlags}"${resolvingText} at "${formatWarningLocation(location)}". Using any instead.`,
-		filePath: location.filePath,
-		line: location.line,
-		column: location.column,
-		parsedSymbolStack: [...context.parsedSymbolStack],
-		typeFlags,
-		typeText,
-		sourceText,
-	});
-}
-
-interface WarningLocation {
-	filePath: string;
-	line: number;
-	column: number;
-	node: ts.Node;
-}
-
-function getWarningLocation(
-	type: ts.Type,
-	typeNode: ts.TypeNode | undefined,
-	context: ParserContext,
-): WarningLocation {
-	const node = getWarningNode(type, typeNode, context) ?? context.sourceFile;
-	const sourceFile = node.getSourceFile();
-	const start = node.getStart(sourceFile);
-	const { line, character } = sourceFile.getLineAndCharacterOfPosition(start);
-
-	return {
-		filePath: sourceFile.fileName,
-		line: line + 1,
-		column: character + 1,
-		node,
-	};
-}
-
-function getWarningNode(
-	type: ts.Type,
-	typeNode: ts.TypeNode | undefined,
-	context: ParserContext,
-): ts.Node | undefined {
-	return (
-		getSymbolDeclaration(type.aliasSymbol) ??
-		getSymbolDeclaration(type.getSymbol()) ??
-		typeNode ??
-		getCurrentSourceTypeNode(context) ??
-		getSubstitutionTypeDeclaration(type) ??
-		getCurrentSourceNode(context)
-	);
-}
-
-function getSubstitutionTypeDeclaration(type: ts.Type): ts.Declaration | undefined {
-	if (!hasExactFlag(type, ts.TypeFlags.Substitution)) {
-		return undefined;
-	}
-
-	const substitutionType = type as ts.SubstitutionType;
-	return (
-		getSymbolDeclaration(substitutionType.baseType.aliasSymbol) ??
-		getSymbolDeclaration(substitutionType.baseType.getSymbol()) ??
-		getSymbolDeclaration(substitutionType.constraint.aliasSymbol) ??
-		getSymbolDeclaration(substitutionType.constraint.getSymbol())
-	);
-}
-
-function getCurrentSourceNode(context: ParserContext): ts.Node | undefined {
-	return context.sourceNodeStack.at(-1);
-}
-
-function getCurrentSourceTypeNode(context: ParserContext): ts.TypeNode | undefined {
-	return getTypeNodeFromNode(getCurrentSourceNode(context));
-}
-
-function getTypeNodeFromNode(node: ts.Node | undefined): ts.TypeNode | undefined {
-	if (!node) {
-		return undefined;
-	}
-	if (ts.isTypeNode(node)) {
-		return node;
-	}
-	if (ts.isTypeAliasDeclaration(node)) {
-		return node.type;
-	}
-	if (ts.isParameter(node) || ts.isPropertySignature(node) || ts.isPropertyDeclaration(node)) {
-		return node.type;
-	}
-	if (
-		ts.isFunctionDeclaration(node) ||
-		ts.isFunctionExpression(node) ||
-		ts.isMethodSignature(node) ||
-		ts.isMethodDeclaration(node)
-	) {
-		return node.type;
-	}
-
-	return undefined;
-}
-
-function formatWarningLocation(location: WarningLocation): string {
-	return `${location.filePath}:${location.line}:${location.column}`;
-}
-
-function getWarningTypeText(type: ts.Type, context: ParserContext): string {
-	return truncateWarningText(
-		context.checker.typeToString(
-			type,
-			undefined,
-			ts.TypeFormatFlags.NoTruncation | ts.TypeFormatFlags.InTypeAlias,
-		),
-	);
-}
-
-function getWarningSourceText(node: ts.Node | undefined): string | undefined {
-	const sourceText = getTypeTextFromNode(node);
-	if (!sourceText) {
-		return undefined;
-	}
-
-	return truncateWarningText(sourceText.replace(/\s+/g, ' ').trim());
-}
-
-function getTypeTextFromNode(node: ts.Node | undefined): string | undefined {
-	return getTypeNodeFromNode(node)?.getText();
-}
-
-function truncateWarningText(text: string): string {
-	const maxLength = 160;
-	return text.length > maxLength ? `${text.slice(0, maxLength - 1)}…` : text;
-}
-
-function getSymbolDeclaration(symbol: ts.Symbol | undefined): ts.Declaration | undefined {
-	return symbol?.declarations?.[0];
-}
-
-const typeFlagDisplayOrder: Array<[ts.TypeFlags, string]> = [
-	[ts.TypeFlags.Any, 'Any'],
-	[ts.TypeFlags.Unknown, 'Unknown'],
-	[ts.TypeFlags.Undefined, 'Undefined'],
-	[ts.TypeFlags.Null, 'Null'],
-	[ts.TypeFlags.Void, 'Void'],
-	[ts.TypeFlags.String, 'String'],
-	[ts.TypeFlags.Number, 'Number'],
-	[ts.TypeFlags.BigInt, 'BigInt'],
-	[ts.TypeFlags.Boolean, 'Boolean'],
-	[ts.TypeFlags.ESSymbol, 'ESSymbol'],
-	[ts.TypeFlags.StringLiteral, 'StringLiteral'],
-	[ts.TypeFlags.NumberLiteral, 'NumberLiteral'],
-	[ts.TypeFlags.BigIntLiteral, 'BigIntLiteral'],
-	[ts.TypeFlags.BooleanLiteral, 'BooleanLiteral'],
-	[ts.TypeFlags.UniqueESSymbol, 'UniqueESSymbol'],
-	[ts.TypeFlags.EnumLiteral, 'EnumLiteral'],
-	[ts.TypeFlags.Enum, 'Enum'],
-	[ts.TypeFlags.NonPrimitive, 'NonPrimitive'],
-	[ts.TypeFlags.Never, 'Never'],
-	[ts.TypeFlags.TypeParameter, 'TypeParameter'],
-	[ts.TypeFlags.Object, 'Object'],
-	[ts.TypeFlags.Index, 'Index'],
-	[ts.TypeFlags.TemplateLiteral, 'TemplateLiteral'],
-	[ts.TypeFlags.StringMapping, 'StringMapping'],
-	[ts.TypeFlags.Substitution, 'Substitution'],
-	[ts.TypeFlags.IndexedAccess, 'IndexedAccess'],
-	[ts.TypeFlags.Conditional, 'Conditional'],
-	[ts.TypeFlags.Union, 'Union'],
-	[ts.TypeFlags.Intersection, 'Intersection'],
-];
-
-function getTypeFlagNames(flags: ts.TypeFlags): string[] {
-	const names = typeFlagDisplayOrder
-		.filter(([flag]) => (flags & flag) === flag)
-		.map(([, name]) => name);
-
-	if (names.length > 0) {
-		return names;
-	}
-
-	return [ts.TypeFlags[flags] ?? String(flags)];
-}
-
-function getTypeId(type: ts.Type): number | undefined {
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	return (type as any).id;
-}
-
-/**
- * Creates a shallow version of a type for cycle detection.
- * Returns the appropriate type node based on the type's structure, but without resolving nested types.
- */
-function createShallowType(
-	type: ts.Type,
-	typeName: TypeName | undefined,
-	checker: ts.TypeChecker,
-): AnyType {
-	// Check for union types
-	if (type.isUnion()) {
-		return new UnionNode(typeName, []);
-	}
-
-	// Check for intersection types
-	if (type.isIntersection()) {
-		return new IntersectionNode(typeName, [], []);
-	}
-
-	// Check for array types
-	if (checker.isArrayType(type)) {
-		// Return an array with 'any' as element type to avoid recursion
-		return new ArrayNode(
-			type.aliasSymbol?.name
-				? new TypeName(type.aliasSymbol.name, typeName?.namespaces, typeName?.typeArguments)
-				: undefined,
-			new IntrinsicNode('any'),
-		);
-	}
-
-	// Check for tuple types
-	if (checker.isTupleType(type)) {
-		return new TupleNode(typeName, []);
-	}
-
-	// Check for function types
-	const callSignatures = type.getCallSignatures();
-	if (callSignatures.length >= 1) {
-		return new FunctionNode(typeName, []);
-	}
-
-	// Default to object type (interfaces, type aliases, classes, etc.)
-	return new ObjectNode(typeName, [], undefined);
-}
-
-// Internal API
-declare module 'typescript' {
-	interface Symbol {
-		parent?: ts.Symbol;
-	}
+	return new TypeResolutionSession(context).resolve(type, typeNode);
 }
