+++ src/parsers/typeResolvers/tupleTypeResolver.ts
@@ -1,0 +1,23 @@
+import ts from 'typescript';
+import { TupleNode, type AnyType } from '../../models';
+import { type TypeResolutionRequest, type TypeResolutionSession } from '../typeResolutionTypes';
+
+// Tuple handling stays separate from arrays because TypeScript
+// exposes tuple element types through tuple-specific metadata and the output
+// model preserves tuple arity.
+
+export function resolveTupleType(
+	{ type, typeName }: TypeResolutionRequest,
+	session: TypeResolutionSession,
+): AnyType | undefined {
+	const { checker } = session.context;
+
+	if (!checker.isTupleType(type)) {
+		return undefined;
+	}
+
+	return new TupleNode(
+		typeName,
+		(type as ts.TupleType).typeArguments?.map((x) => session.resolve(x, undefined)) ?? [],
+	);
+}
